Move code to the new assembly
