+++ src/Integration.Vsix/AsmRef_Integration.Vsix_Baseline_WithStrongNames.txt
@@ -1,7 +1,7 @@
 ﻿---
 ################################
 # Assembly references report
-# Report date/time: 2022-03-01T10:01:54.2315528Z
+# Report date/time: 2022-03-02T11:47:16.0039792Z
 ################################
 #
 # Generated by Devtility CheckAsmRefs v0.11.0.223
@@ -19,7 +19,7 @@
 #
 ################################
 
-# Base directory: c:\Projects\sonarlint-visualstudio\src\Integration.Vsix\bin\Debug\VS2022\net472
+# Base directory: C:\Projects\sonarlint-visualstudio\src\Integration.Vsix\bin\Debug\VS2022\net472
 Include patterns:
 - 'Sonar*'
 Exclude patterns:
@@ -117,6 +117,7 @@ Referenced assemblies:
 - 'SonarLint.VisualStudio.Infrastructure.VS, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'SonarLint.VisualStudio.Integration, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'SonarLint.VisualStudio.IssueVisualization, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
+- 'SonarLint.VisualStudio.Roslyn.Suppressions, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'SonarQube.Client, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'System.ComponentModel.Composition, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
@@ -124,7 +125,7 @@ Referenced assemblies:
 - 'System.IO.Abstractions, Version=7.0.0.0, Culture=neutral, PublicKeyToken=96bf224d23c43e59'
 - 'System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'System.Xaml, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
-# Number of references: 30
+# Number of references: 31
 
 ---
 Assembly: 'SonarLint.VisualStudio.CFamily, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
@@ -338,7 +339,10 @@ Relative path: 'SonarLint.VisualStudio.Roslyn.Suppressions.dll'
 
 Referenced assemblies:
 - 'mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
-# Number of references: 1
+- 'SonarLint.VisualStudio.Core, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
+- 'SonarQube.Client, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
+- 'System.ComponentModel.Composition, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
+# Number of references: 4
 
 ---
 Assembly: 'SonarLint.VisualStudio.TypeScript, Version=6.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
