+++ src/Integration.Vsix/SonarLintIntegrationPackage.cs
@@ -29,8 +29,8 @@ using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Core.Suppression;
-using SonarLint.VisualStudio.Integration.Suppression;
 using SonarLint.VisualStudio.Integration.TeamExplorer;
+using SonarLint.VisualStudio.Roslyn.Suppressions.InProcess;
 
 namespace SonarLint.VisualStudio.Integration.Vsix
 {
